+++ assembler/src/lexer/tokenizer.rs
@@ -1227,13 +1227,13 @@ impl<'a> Tokenizer<'a> {
     pub fn curr_str(&self) -> String {
         if self.matching {
             String::from_utf8_lossy(
-                &self.bytes[self.start_curr.index_real as usize
-                    ..(self.current.index_real - self.c.len_utf8()) as usize],
+                &self.bytes[self.start_curr.index_real
+                    ..(self.current.index_real - self.c.len_utf8())],
             )
             .to_string()
         } else {
             String::from_utf8_lossy(
-                &self.bytes[self.start_curr.index_real as usize..self.current.index_real as usize],
+                &self.bytes[self.start_curr.index_real..self.current.index_real],
             )
             .to_string()
         }
